Charles G. Waldman <cgw@fnal.gov>: Add the EXTENDED_ARG opcode to the virtual machine, allowing 32-bit arguments to opcodes instead of being forced to stick to the 16-bit limit.  This is especially useful for machine-generated code, which can be too long for the SET_LINENO parameter to fit into 16 bits.
This closes the implementation portion of SourceForge patch #100893.

+++ Include/node.h
@@ -19,8 +19,8 @@ extern "C" {
 typedef struct _node {
     short		n_type;
     char		*n_str;
-    short		n_lineno;
-    short		n_nchildren;
+    int			n_lineno;
+    int			n_nchildren;
     struct _node	*n_child;
 } node;
